feat(home): Specs and Tasks become widgets
Move the two planning cards behind the contract, rendering unchanged. The
board no longer knows what is in the Project planning group — it asks the
registry, mounts what comes back, and routes each source to the widgets
that declared it.

The consequence is the point: laneBoard.js has no ipcRenderer listeners
left. SPEC_DATA, TASKS_DATA, GIT_STATUS_DATA, the dispatch activity hooks,
_fetchForProject and _refreshData all live in homeData now, and the electron
and ipcChannels imports left with them. One subscription per source, however
many widgets read it.

_card() is gone too — _buildTerminalsCard calls widgetShell(), so the shell
has one definition rather than two until T05 retires the card.

T03 · spec:home-widget-board
